Working DRG updated and cleaned up.
